requirements.txt
langchain
langchain_community
langchain_core
python-dotenv
langchain_ollama
streamlit
